// backend/src/models/types/user-type.ts
export default interface UserType{

			firstname: string
			lastname: string
			middlename: string
			avatar: string
			signature: string
			authid: number
			email: string
			phone_ext: number
			phone_number: number
			phone_verified: string
			allowphonedisplay: number
			default_language: number
			default_currency: number
			currentcompanyid: number
			role: number

}